requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
frozendict==2.4.5
fsspec==2024.9.0
gradio==5.0.2
gradio_client==1.4.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
multitasking==0.0.11
nltk==3.9.1
numpy==2.1.2
oauthlib==3.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.6
praw==7.7.1
prawcore==2.4.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.2
rsa==4.9
ruff==0.6.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
Telethon==1.37.0
threadpoolctl==3.5.0
tomlkit==0.12.0
tqdm==4.66.5
tweepy==4.14.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.2.3
uvicorn==0.31.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
yfinance==0.2.44
